Added number function file; unit conversion
A general, scalable approach to converting units.

Issue with conversion and floating point meant some values were just
slightly off. Added precision rounding, default to 4 decimal places to
deal with the issue.

Sass in memory will keep numbers up to however many decimals, but will
not output any more than 4 in the language, so it may as well just be
rounded. This causes issues in comparing values, and made it impossible
to know if a number was really the number it shows it as.