Fixed bug with factor and base font size conversion
What was happening with a relative base font size, is that the factor
conversion was taking place twice if no factor was passed. So the
default factor is the font size, and what happens is that the font size
gets normalized, then the factor gets converted based on the font size.
This meant that in the event of no passed factor, the font size was
getting converted against... the font size. Which is not what we want to
happen by default. HOWEVER this is expected behaviour when passing a
factor. So set a default to something other than the font size to get
around this problem.