Fixed color method for libsass, added font methods
The issue in libsass in version 3.2 is that quote-wrapped strings that
are also css colors (like "white") aren't correctly treated like
strings, but as colors. This means that the type check for string will
not pass for things like "white" or "black". The method now is very
liberal in what it accepts to get around this.