Renamed so they don't clash when loading the path
When creating the grunt file, best practice is to include the folder to
hydro-css so you don't have to reference the whole location in your Sass
code. That means any files in this scss/ dir would be accessible by just
it's file name (@import 'hydro'). Added the hydro prefix so 'core' and
'theme' won't clash with anything else and can be safely referenced.